agents/google-trends-agent/pyproject.toml
[project]
name = "google-trends-agent"
version = "0.1.0"
description = "AI Agent that is designed to surface the newest Google Trends in realtime"
authors = [
    {name = "Yehia Elshater",email = "elshater@google.com"}
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines] (>=1.104.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-genai (>=1.26.0,<2.0.0)",
    "google-cloud-bigquery (>=3.35.0,<4.0.0)",
    "google-adk (>=1.7.0,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.35.0"
jinja2 = "^3.1.6"
python-dotenv = "^1.1.1"
google-genai = "^1.26.0"
google-cloud-bigquery = "^3.35.0"
google-adk = "^1.7.0"
pandas = "^2.3.1"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.2"
google-cloud-aiplatform = { version = ">=1.104.0", extras = ["adk","agent-engines"] }


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
